fix bugs in register
